+++ src/services/LIFI/executionFiles/allowance.execute.ts
@@ -1,9 +1,9 @@
 import { JsonRpcSigner } from '@ethersproject/providers'
-import { Chain, Execution, Token } from '@lifinance/types'
 import BigNumber from 'bignumber.js'
 import { constants } from 'ethers'
 
 import { createAndPushProcess, setStatusDone, setStatusFailed } from '../status'
+import { Chain, Execution, Token } from '../types'
 import { getApproved, setApproval } from '../utils'
 
 export const checkAllowance = async (
